requirements.txt
Django==5.0.1
# django-cors-headers==4.3.1
# django-crispy-forms==2.1
# django-taggit==5.0.1
djangorestframework==3.14.0
# psycopg2-binary==2.9.9
# python-dotenv==1.0.1
selenium==4.17.2
# crispy-bootstrap5==2023.10
channels==4.0.0
daphne==4.0.0
graphene-django==3.2.0